Reorganized XLSForm export tests and fixtures.
