+++ pyxform/tests/example_xforms/single_select_one_survey.xml
